+++ foundation/eclipselink.core.test/src/it/java/org/eclipse/persistence/testing/tests/distributedservers/rcm/ConfigurableCacheSyncDistributedTest.java
@@ -126,8 +126,8 @@ public class ConfigurableCacheSyncDistributedTest extends AutoVerifyTestCase {
                 int cacheSyncType = descriptor.getCacheSynchronizationType();
                 Object newCacheSyncType = cacheSyncConfigValues.get(keyClass);
                 if (newCacheSyncType != null) {
-                    oldCacheSyncConfigValues.put(keyClass, Integer.valueOf(cacheSyncType));
-                    descriptor.setCacheSynchronizationType(((Integer)newCacheSyncType).intValue());
+                    oldCacheSyncConfigValues.put(keyClass, cacheSyncType);
+                    descriptor.setCacheSynchronizationType((Integer) newCacheSyncType);
                 }
             }
         }
@@ -150,7 +150,7 @@ public class ConfigurableCacheSyncDistributedTest extends AutoVerifyTestCase {
         while (keys.hasMoreElements()) {
             Class keyClass = (Class)keys.nextElement();
             ClassDescriptor descriptor = getSession().getDescriptor(keyClass);
-            int newCacheSyncType = ((Integer)oldCacheSyncConfigValues.get(keyClass)).intValue();
+            int newCacheSyncType = (Integer) oldCacheSyncConfigValues.get(keyClass);
             descriptor.setCacheSynchronizationType(newCacheSyncType);
         }
     }
