+++ foundation/org.eclipse.persistence.oracle.test/src/it/java/org/eclipse/persistence/testing/tests/lob/ImageSimulator.java
@@ -73,7 +73,7 @@ public class ImageSimulator {
         new Random().nextBytes(pictures);
         Byte[] pics = new Byte[cycle];
         for (int x = 0; x < cycle; x++) {
-            pics[x] = Byte.valueOf(pictures[x]);
+            pics[x] = pictures[x];
         }
         return pics;
     }
